Move all resolvers to sqlite specific directory
