+++ src/query/resolvers/sqlite/groupId.js
@@ -1,5 +1,5 @@
-const _ = require('../../utils');
-const debug = require('../../debugger');
+const _ = require('../../../utils');
+const debug = require('../../../debugger');
 
 // This resolver will add the json_group_array property into the query select area. The
 // grouped ids can be used by the next nested nodes when its parent data is grouped.
