+++ src/query/resolvers/sqlite/main.js
@@ -1,5 +1,5 @@
-const _ = require('../../utils');
-const debug = require('../../debugger');
+const _ = require('../../../utils');
+const debug = require('../../../debugger');
 
 module.exports = function graphNodeResolver(node, options = {}, nextNodes, customResolver, parentNode) {
